--- packages/client/src/pages/Forum/index.tsx
@@ -1,1 +1,0 @@
-export { ForumPage } from '../Forum/ForumPage';
